boost query flexibility
the distinction between entity and property queries is gone, it's now possible to retrieve entities and/or properties in one query. if only one entity/property is queried for, the result is a flat array. if the query contains multiple selections, an array of objects is returned. the property names can be defined in the query using "as [propname]".

all selected columns are now aliased in the final sql statement - this was necessary to allow aggressive storable selection in combination with property/aggregation selections. based on the selectclause in the query's AST a set of jdbc resultset collectors is created (by collectorgenerator visitor). each single collector extracts the values from the resultset and return them or creates the appropriate storable.