modified storable behaviour to avoid copying of primitive property values
the internal _props object now only contains modified primitive values and object/collection mappings. this way it's not necessary to copy all primitive values of a storable from its _entity to _props object when accessing a property for the first time (which can be quite expensive in both terms of performance and memory consumption).