moved requiring of ast module out of  Parser method.
since a recent commit in ringojs (https://github.com/ringo/ringojs/commit/26fd4159e417c3f0cdaa91a3c285e675fdb6822a) this lead to problems if the store is instantiated as a singleton (which it most likely is), notably instanceof ast nodes doesn't work anymore in collectorgenerator.

the downside of this change is that parser.js needs some manual work when regenerated.