modified behaviour of storables reg. undefined values during construction time
if a storable is instantiated with undefined properties sqlstore now returns null
for them on access. before this fix a storable would return `undefined` for properties
that haven't been defined in the properties hash during instantiation, but return
`null` once the storable is fetched from db or cache. this is consistent with the
behaviour of storables loaded freshly from db.