allow loading of entity modules
store now provides a registerEntityModule() method to register module paths
that define entities. this is especially necessary if entities are defined in
separate modules, and module reloading is enabled: in such cases
the store might be re-initialized with only a subset or no entity constructors
at all, leading to "entity not registered" exceptions. With this change
sqlstore can now load entity modules on demand.