fixed expression parsing:
a single AND condition is now merged into OR conditions, if present. this way i.e.
`Foo.x = y or Foo.x = z` leads to an expression instance without AND conditions
(before both would be split into single AND/OR condition lists).

The ConditionList constructor now takes the logical operator as third argument,
this was hardcoded in sqlgenerator (to AND) before.

fixes #31