various major changes, most notable:
- it's no longer needed to prefix every part of the select clause with either
an alias or the entity name. sql query generation now starts with the from and
join clauses and keeps track of the referenced entities, and from then resolves
non-qualified Idents (eg. "title" instead of "Book.title" or "b.title") to the
appropriate mapping.
- the query result collector function is now created by the sql generator, which
makes it possible to cache the collector as well instead of creating the collector
for every issued sql query. collections however still use their own collectors,
these both will possibly be unified in the nearer future.

other changes:
- simplified collectors: they no longer store the column offset themselves,
but get it passed as argument from the wrapping collector function
- added a bunch of different tests